# openapi.yaml
openapi: 3.1.0
info:
  title: Tateca API
  description: "Group expense management API with multi-currency support. Track expenses,\
    \ loans, and settlements across multiple users and currencies."
  contact:
    name: Tateca Team
    url: https://tateca.net
    email: support@tateca.net
  license:
    name: Proprietary
    url: https://tateca.net/terms
  version: 1.0.0
...
servers:
- url: http://localhost:8080
  description: Local Development
- url: https://staging-api.tateca.net
  description: Staging
- url: https://api.tateca.net
  description: Production
security:
- bearerAuth: []
...
paths:
...
  /auth/users:
    post:
      tags:
      - Auth Users
      summary: Create new authenticated user
      description: Registers a new authenticated user in the system
      operationId: createAuthUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAuthUserRequestDTO"
        required: true
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponseDTO"
        "401":
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              example:
                status: 401
                message: Unauthorized
        "409":
          description: Conflict - User already exists
          content:
            application/json:
              example:
                status: 409
                message: Email already exists
        "500":
          description: Internal Server Error
          content:
            application/json:
              example:
                status: 500
                message: Database error
        "400":
          description: Validation error - Invalid request parameters
          content:
            application/json:
              examples:
                Email is null, empty, or whitespace only:
                  description: "When email is null, empty string, or only whitespace"
                  value:
                    status: 400
                    message: "email: Email is required and cannot be blank"
                Email exceeds maximum length:
                  description: When email exceeds 255 characters
                  value:
                    status: 400
                    message: "email: Email must not exceed 255 characters"
                Malformed JSON:
                  description: When request body is not valid JSON
                  value:
                    status: 400
                    message: Invalid request body format. Please check the data types
                      and structure.
        "415":
          description: Unsupported Media Type - Content-Type header must be application/json
          content:
            application/json:
              example:
                status: 415
                message: Unsupported Media Type. Content-Type must be application/json
  /users/{userId}:
    patch:
      tags:
      - Users
      summary: Update user name
      operationId: updateUserName
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserNameRequestDTO"
        required: true
      responses:
        "200":
          description: User name updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "400":
          description: Validation error - Invalid request parameters
          content:
            application/json:
              examples:
                Empty or blank name:
                  description: "When user_name is null, empty string, or only whitespace"
                  value:
                    status: 400
                    message: "name: User name is required and cannot be blank"
                Name exceeds maximum length:
                  description: When user_name exceeds 50 characters
                  value:
                    status: 400
                    message: "name: User name must be between 1 and 50 characters"
                Invalid UUID format:
                  description: When userId path parameter has invalid UUID format
                  value:
                    status: 400
                    message: "Invalid format for parameter 'userId': expected UUID"
                Malformed JSON:
                  description: When request body is not valid JSON
                  value:
                    status: 400
                    message: Invalid request body format. Please check the data types
                      and structure.
        "500":
          description: Internal Server Error
          content:
            application/json:
              example:
                status: 500
                message: An unexpected error occurred
        "415":
          description: Unsupported Media Type - Content-Type header must be application/json
          content:
            application/json:
              example:
                status: 415
                message: Unsupported Media Type. Content-Type must be application/json
        "404":
          description: User not found
          content:
            application/json:
              example:
                status: 404
                message: User not found
        "401":
          description: Unauthorized - Invalid or missing Firebase JWT token
          content:
            application/json:
              example:
                status: 401
                message: Unauthorized
  /auth/users/review-preferences:
    patch:
      tags:
      - Auth Users
      summary: Update app review preferences
      description: Updates the app review preferences for an authenticated user
      operationId: updateReviewPreferences
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppReviewRequestDTO"
        required: true
      responses:
        "200":
          description: Review preferences updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponseDTO"
        "401":
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              example:
                status: 401
                message: Unauthorized
        "500":
          description: Internal Server Error
          content:
            application/json:
              example:
                status: 500
                message: Database error
        "404":
          description: User not found - The specified UID does not exist
          content:
            application/json:
              example:
                status: 404
                message: User not found
        "415":
          description: Unsupported Media Type - Content-Type header must be application/json
          content:
            application/json:
              example:
                status: 415
                message: Unsupported Media Type. Content-Type must be application/json
        "400":
          description: Validation error - Invalid request parameters
          content:
            application/json:
              examples:
                App review status is null:
                  description: When app_review_status is null
                  value:
                    status: 400
                    message: "appReviewStatus: App review status is required"
                Invalid ENUM value:
                  description: "When app_review_status is not a valid enum value (PENDING,\
                    \ COMPLETED, PERMANENTLY_DECLINED)"
                  value:
                    status: 400
                    message: "Invalid value for app_review_status: must be PENDING,\
                      \ COMPLETED, or PERMANENTLY_DECLINED"
                Malformed JSON:
                  description: When request body is not valid JSON
                  value:
                    status: 400
                    message: Invalid request body format. Please check the data types
                      and structure.
...
  /auth/users/{uid}:
    get:
      tags:
      - Auth Users
      summary: Get authenticated user information
      description: Retrieves detailed information about an authenticated user by their
        UID
      operationId: getAuthUser
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
          maxLength: 128
          minLength: 0
      responses:
        "200":
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponseDTO"
        "400":
          description: Validation error - Invalid UID format
          content:
            application/json:
              example:
                status: 400
                message: Invalid UID format
        "404":
          description: User not found - The specified UID does not exist
          content:
            application/json:
              example:
                status: 404
                message: User not found
    delete:
      tags:
      - Auth Users
      summary: Delete authenticated user
      description: Removes an authenticated user from the system
      operationId: deleteUserInfo
      parameters:
      - name: uid
        in: path
        required: true
        schema:
          type: string
          maxLength: 128
          minLength: 0
      responses:
        "400":
          description: Validation error - Invalid UID format
          content:
            application/json:
              example:
                status: 400
                message: Invalid UID format
        "404":
          description: User not found - The specified UID does not exist
          content:
            application/json:
              example:
                status: 404
                message: User not found
        "204":
          description: User deleted successfully
...
  /groups/{groupId}/users/{userUuid}:
    delete:
      tags:
      - Groups
      summary: Leave a group
      operationId: leaveGroup
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: userUuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: User left group successfully
        "404":
          description: Group not found
          content:
            application/json:
              example:
                status: 404
                message: Group not found
        "400":
          description: Validation error - Invalid UUID format
          content:
            application/json:
              examples:
                Invalid groupId UUID format:
                  description: When groupId path parameter has invalid UUID format
                  value:
                    status: 400
                    message: "Invalid format for parameter 'groupId': expected UUID"
                Invalid userUuid UUID format:
                  description: When userUuid path parameter has invalid UUID format
                  value:
                    status: 400
                    message: "Invalid format for parameter 'userUuid': expected UUID"
        "403":
          description: Forbidden - User doesn't have permission to remove this user
          content:
            application/json:
              example:
                status: 403
                message: Forbidden
        "401":
          description: Unauthorized - Invalid or missing Firebase JWT token
          content:
            application/json:
              example:
                status: 401
                message: Unauthorized
components:
  schemas:
...
    AuthUserResponseDTO:
      type: object
      description: Authentication user information
      properties:
        uid:
          type: string
          description: Firebase authentication user ID
          example: firebase-uid-123
        name:
          type: string
          description: User's name from authentication provider
          example: John Doe
        email:
          type: string
          description: User's email address
          example: john.doe@example.com
        created_at:
          type: string
          description: Account creation timestamp (Tokyo time)
          example: 2024-01-01T12:00:00+09:00
        updated_at:
          type: string
          description: Last update timestamp (Tokyo time)
          example: 2024-01-15T14:30:00+09:00
        last_login_time:
          type: string
          description: Last login timestamp (Tokyo time)
          example: 2024-01-20T10:15:00+09:00
        total_login_count:
          type: integer
          format: int32
          description: Total number of logins
          example: 42
        last_app_review_dialog_shown_at:
          type: string
          description: Timestamp when app review dialog was last shown (Tokyo time)
          example: 2024-01-18T16:45:00+09:00
        app_review_status:
          type: string
          description: Current app review status
          enum:
          - PENDING
          - COMPLETED
          - PERMANENTLY_DECLINED
...
    UserResponseDTO:
      type: object
      description: User information response
      properties:
        uuid:
          type: string
          description: User's unique identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          description: User's display name
          example: John Doe
        auth_user:
          $ref: "#/components/schemas/AuthUserResponseDTO"
          description: Authentication user details
        created_at:
          type: string
          description: User creation timestamp (Tokyo time)
          example: 2024-01-01T12:00:00+09:00
        updated_at:
          type: string
          description: User last update timestamp (Tokyo time)
          example: 2024-01-15T14:30:00+09:00
...
    CreateAuthUserRequestDTO:
      type: object
      description: Request to create a new authenticated user
      properties:
        email:
          type: string
          description: User's email address
          example: john@example.com
          maxLength: 255
          minLength: 0
      required:
      - email
    UpdateUserNameRequestDTO:
      type: object
      description: Request to update user name
      properties:
        user_name:
          type: string
          description: User's display name
          example: John Doe
          maxLength: 50
          minLength: 1
      required:
      - user_name
...
    UpdateAppReviewRequestDTO:
      type: object
      description: Request to update app review preferences
      properties:
        app_review_status:
          type: string
          description: App review status
          enum:
          - PENDING
          - COMPLETED
          - PERMANENTLY_DECLINED
          example: COMPLETED
      required:
      - app_review_status
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Firebase JWT token for user authentication
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      description: API Key for internal system operations (EventBridge + Lambda)
      name: X-API-Key
      in: header